requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
distro==1.9.0
duckdb==1.2.1
et_xmlfile==2.0.0
Faker==19.13.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.31.0
numpy==1.24.4
openai==1.67.0
openpyxl==3.1.5
packaging==24.2
pandas==1.5.3
pandasai==2.2.15
pillow==10.4.0
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.1
SQLAlchemy==2.0.39
sqlglot==25.34.1
sqlglotrs==0.3.0
streamlit==1.43.2
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
